Fixing SUSE/OpenSUSE platform specific settings
SUSE platform support was broken, because:
 - SUSE no longer uses /etc/httpd as config dir
 - Calls service apache2 instead of httpd
 - Creates binary/pidfile with using name `httpd2`
 - Uses custom user / group to run Apache
 - Installs a2enmod / a2dismod what works differently

The last point needs tidying up, I currently remove these files and
install cookbook's one, but it is not prevents replacing
these scripts with the upgrade of apache2 package (what provides them).
The main problem is a2dismod is a symlink to a2enmod, and it is
confuses template resource. This is why I have to explicitly delete
these files