templatize the envvars file
The envvars file (typically installed into /etc/apache2/envvars) sets some environment variables that are used by the init script.

It sets the user to run as, group, pidfile location, etc.

Previous versions of the cookbook didn't include this file, so the init system would be using the wrong pidfile location, for example, and wouldn't know how to stop/restart apache.

This version uses the same values that are used in the rest of the apache2 configuration.