Allow to set LANG and LC_ALL in envvars
Some apps/modules like mod_dav and mod_wsgi need the language to be set to specific values. Lets allow that through envvars (on debian systems only?).

This should be backward-compatible as 'C' remains the default. If a language is set, it is used, if the parameter is set to 'system', /etc/default/locale gets importet.