FCGI paths should not be messed with on RHEL/CenOS 7. CentOS 7 (and recent Fedoras) have Apache 2.4, where FCGI socket path and shared memory path is managed adequately without further involvment neccessary (subdirectory is created under /var/run/httpd).
However, when the path is specified manually, that subdirectory is NOT
created automatically if it doesn't already exist and Apache fails to
start. Since in recent RHEL systems /var/run is on tmpfs, end result is
that chef-created subdirectory disappears on shutdown, and Apache fails
to start after server reboot. Best to leave out these two settings
unset then.