Fixes #317 Import default recipe into all definitions
If the apache2::default recipe is not included in definitions where
notifications are raised against the apache2 service then an error can
occur becuase the service is not defined. That forces the resonsibility
for including the default recipe onto the user. By including the
default recipe into all the definitions they become more fault tolerant
and so more useable.