Use an array as listen attribute instead of a hash
This PR amends the logic introduced in
https://github.com/svanzoest-cookbooks/apache2/pull/378 to use an array
as attribute instead of a hash, as this will simplify using this
cookbook in other recipes.

E.g., if a recipe wants httpd to listen on `addr:port`, it is much easier to
just set

    node.set['apache2']['listen'] += ['addr:port']

instead of first having to check whether the listen hash does have the
key `addr` already and then either add it or update the port list.