change listen attributes to be hash of addresses and ports to listen to
This should fix the issue that was going on w/ PR #372
The attributes `node['apache']['listen']` will be a hash indexed by address where the values are ports to listen to.

For example:

```json
{
  "apache": {
    "listen": {
      "*": ["80", "443"],
      "127.0.0.1": [ "12345" ]
    }
  }
}
```

would create this ports.conf:

```
Listen *:80
Listen *:443
Listen 127.0.0.1:12345
```

Which was impossible with the older listen_ports and listen_addresses attributes.