Fix issue when using apache2_mod_ssl resource instead a wrapper resource
When doing `apache2_module 'ssl'` instead of another custom resource, the
mod_ssl package does not get installed properly due to the fact it's set to run
in the root context of the chef run. However, it never seems to get installed,
or at least installed in time for it to work leaving a system in a broken state.

I cannot see why this was needed in the first place as it seems to work without
it.

Signed-off-by: Lance Albertson <lance@osuosl.org>