Match lock directory permissions and ownership to distribution for RHEL
This primarily fixes the permissions on RHEL-based systems to match what they
have by default. In some situations, this was leading to an idempotency issues
on RHEL systems. To work around this, setting the correct permission and
ownership works around this from happening.

It appears this permission/ownership was taken from a Debian-based system which
is why this was happening.

Signed-off-by: Lance Albertson <lance@osuosl.org>